refactor: 자동차 생성 기능 클래스 분리
